+++ NEWS.md
@@ -23,6 +23,7 @@
 
 ## Bug fixes
 
+- `effectsize()` for `htest` works with Spearman and Kendall correlations ( #165 ).  
 - `cramers_v()` and `phi()` now work with goodness-of-fit data ( #158 )
 - `standardize_parameters()` for post-hoc correctly standardizes transformed outcome.
 - Setting `two_sd = TRUE` in `standardize()` and `standardize_parameters()` (correctly) on uses 2-SDs of the predictors (and not the response).